pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
[tool.poetry.dependencies]
curl-cffi = "^0.14.0"
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.135.1"}
httpx = "^0.28.1"
beautifulsoup4 = "^4.14.3"
lxml = "^6.0.2"
python-dotenv = "^1.2.2"
asyncpg = "^0.30.0"
bcrypt = "^5.0.0"
PyJWT = "^2.9.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
